+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Remove unnecessary multi_x convenience functions
 - Refactor so that API functions are fully self-contained
+- Add output data to input CSV file
 
 ## [0.0.3] - 2026-06-11